Split lowering of declarations
